+++ src/UserCollection.java
@@ -3,6 +3,6 @@ import java.util.List;
 public interface UserCollection {
     void displayUsers();
     void addUser(User user);
-    User createUser(String name, Integer contact, String address);
+    User createUser(String name, long contact, String address);
     Book searchUserById(Integer id);
 }
